% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atraso_contrato.R
\name{atraso_contrato}
\alias{atraso_contrato}
\title{Multa por Atraso Contrato Original}
\usage{
atraso_contrato(num_meses_atraso_contrato = 18)
}
\arguments{
\item{num_meses_atraso_contrato}{A number}
}
\value{
total payment for the informed number of months delayed
}
\description{
Computes the fine due to delay in construction, as per in original contract
}
\examples{
 atraso_contrato(10)

}
