% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atraso_contrato_phase1.R
\name{atraso_contrato_phase1}
\alias{atraso_contrato_phase1}
\title{Multa da fase 1 por Atraso Contrato Original}
\usage{
atraso_contrato_phase1(num_meses_atraso_phase1 = 18)
}
\arguments{
\item{num_meses_atraso_phase1}{Number of months delayed in phase 1 (there is no specifics about blocks/chunks)}
}
\value{
total payment for the informed number of months delayed
}
\description{
Computes the fine due to delay in construction, as per in original contract
}
\examples{
 atraso_contrato_phase1(18)

}
