% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atraso_contrato_phase2.R
\name{atraso_contrato_phase2}
\alias{atraso_contrato_phase2}
\title{Multa por Atraso Contrato Original da fase 2}
\usage{
atraso_contrato_phase2(num_meses_atraso_phase2 = 12,
  num_meses_atraso_phase2_equipamento = 12)
}
\arguments{
\item{num_meses_atraso_phase2}{Number of months delayed in phase 2 (there is no specifics about blocks/chunks)}

\item{num_meses_atraso_phase2_equipamento}{Number of months delayed in equipments for phase 2 (there is no specifics about blocks/chunks)}
}
\value{
total payment for the informed number of months delayed
}
\description{
Computes the fine due to delay in construction, as per in original contract
}
\examples{
 atraso_contrato_phase2(12, 12)

}
