% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{constant}
\alias{constant}
\title{Simulates a constant time series}
\usage{
constant(start_seed, num_years = 33)
}
\arguments{
\item{start_seed}{A number}

\item{num_years}{A number}
}
\value{
a vector of a constant times series of size equal to num_years
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 constant(start_seed = 1.3, mu = 0, sd = .05)

}
