% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cambio_futuro.R
\name{gen_cambio_futuro}
\alias{gen_cambio_futuro}
\title{Taxa de cambio futura}
\usage{
gen_cambio_futuro(start_year = 2005, amp_erro = 2, mu = 0)
}
\arguments{
\item{start_year}{A number. the year the series starts}

\item{amp_erro}{A number. Amplitude of error}

\item{mu}{mean of normal error}
}
\value{
data frame with real yearly exchange rate and a column of date
}
\description{
gera projecao de taxa de cambio real futura
}
\examples{
 gen_cambio_futuro()

}
