% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fin_lp1.R
\name{gen_fin_lp1}
\alias{gen_fin_lp1}
\title{Financiamento de longo prazo 1}
\usage{
gen_fin_lp1()
}
\value{
a vector financiamento de longo prazo
}
\description{
Funcao auxiliar que gera vetor de financiamento de longo prazo, para computar juros
}
\examples{
 gen_fin_lp1()

}
