% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fin_lp2.R
\name{gen_fin_lp2}
\alias{gen_fin_lp2}
\title{Financiamento de longo prazo 2}
\usage{
gen_fin_lp2()
}
\value{
a vector financiamento de longo prazo
}
\description{
Funcao auxiliar que gera vetor de financiamento de longo prazo, para computar juros
}
\examples{
 gen_fin_lp2()

}
