% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_interest_ep.R
\name{gen_interest_ep}
\alias{gen_interest_ep}
\title{Interest paid in emprestimo ponte}
\usage{
gen_interest_ep()
}
\value{
a vector with interest paid for 'long tearm loan'emprestimo ponte' per period
}
\description{
computes interest paid for emprestimo ponte
}
\examples{
 gen_interest_ep()

}
