% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_interest_lp.R
\name{gen_interest_lp}
\alias{gen_interest_lp}
\title{Interest paid in long term loan}
\usage{
gen_interest_lp(type_lp = 1, num_round = 0)
}
\arguments{
\item{type_lp}{A number, either 1 or 2 (for type of long term loan)}

\item{num_round}{a number to round numbers. If NA, there is no rounding.}
}
\value{
a vector with interest paid for long tearm loan per period
}
\description{
computes interest paid for long term loan
}
\examples{
 gen_interest_lp(type_lp = 2)

}
