% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_price_ticket_line4.R
\name{gen_price_ticket_line4}
\alias{gen_price_ticket_line4}
\title{Gera tarifa da linha 4}
\usage{
gen_price_ticket_line4(num_years, t_0 = 2.14, ipc_0 = 1.1,
  ipgm_0 = 1.1, a = 0.5, b = 0.5, ipc_realizado = NA,
  igpm_realizado = NA, sensibilidade = 1, ajuste_inflacao = F)
}
\arguments{
\item{num_years}{A number}

\item{t_0}{A number}

\item{ipc_0}{A number}

\item{ipgm_0}{A number}

\item{a}{A number}

\item{b}{A number}

\item{ipc_realizado}{A number}

\item{igpm_realizado}{A number}

\item{sensibilidade}{A vector}

\item{ajuste_inflacao}{logical}
}
\value{
a vector of price tickets o size equal to num_years
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 gen_num_passengers(sensibilidade = 1.3)

}
