% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sensibilidade.R
\name{gen_sensibilidade}
\alias{gen_sensibilidade}
\title{Simulates a vector of sensibilidade parameter}
\usage{
gen_sensibilidade(start_seed, num_years = 33, mu_sense = 0,
  sd_sense = 0.05, type = "random_walk")
}
\arguments{
\item{start_seed}{A number}

\item{num_years}{A number}

\item{mu_sense}{A number}

\item{sd_sense}{A number}

\item{type}{A character. Either "random_walk", "white_noise" or "constant"}
}
\value{
a vector of a sensibilidade of size equal to num_years
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 gen_sensibilidade(start_seed = 1.3, mu = 0, sd = .05)

}
