% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ticket_revenue.R
\name{gen_ticket_revenue}
\alias{gen_ticket_revenue}
\title{Computes tariff revenut (receita tarifaria)}
\usage{
gen_ticket_revenue(num_years = 33, t_0 = 2.14, ipc_0 = 1.1,
  ipgm_0 = 1.1, a = 0.5, b = 0.5, ipc_realizado = NA,
  igpm_realizado = NA, qualityAdjustment = 1, sensibilidade = 1,
  ajuste_inflacao, use_random_walk = F, start_value, mu, sd, beta,
  jump = F)
}
\arguments{
\item{num_years}{A number}

\item{t_0}{A number}

\item{ipc_0}{A number}

\item{ipgm_0}{A number}

\item{a}{A number}

\item{b}{A number}

\item{ipc_realizado}{A number}

\item{igpm_realizado}{A number}

\item{qualityAdjustment}{A number}

\item{sensibilidade}{A vector}

\item{ajuste_inflacao}{logical}

\item{use_random_walk}{Logical. If TRUE, it will use a random walk to forecast demand (instead of a fixed forecast as in excel)}

\item{start_value}{A number. First point of the series. Use only if random_walk = T}

\item{mu}{A number}

\item{sd}{A number}

\item{beta}{A number. Effect of past step on the next step.}

\item{jump}{logical. If TRUE, in the eight year of the series the forecasted demand increases by 25\% relative to last value (plus error). Equivalent to set beta =1.25 for year 8.}
}
\value{
a vector of revenue of size equal to num_years - as a result of real demand as set by sensibilidade
}
\description{
Computes total revenue from tariffs. The formula is: Tariff revenue = ((Exclusive passengers x Implicit Tariff) + .5xIntegrated passengers x Implicit tariff)) x (.8 + .1xlqs + .1xlqm). lqs and lqm are set to 1 in this version.
}
\examples{
 gen_num_passengers(sensibilidade = 1.3)

}
