% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_medio_saldo.R
\name{gera_media_saldo}
\alias{gera_media_saldo}
\title{Roll average for balance}
\usage{
gera_media_saldo(saldo)
}
\arguments{
\item{saldo}{A vector of balance data}
}
\value{
mean of balances
}
\description{
Auxuliar function
}
\examples{
 meu_saldo <- gera_saldo_ponte(c(0, 100, 0), c(100, 100, 100))
gera_saldo_ponte(meu_saldo)

}
