% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_saldo_fin.R
\name{gera_saldo_fin}
\alias{gera_saldo_fin}
\title{Saldo para emprestimo de longo prazo}
\usage{
gera_saldo_fin(amortizacao, emprestimo, num_round = NA)
}
\arguments{
\item{amortizacao}{A vector with amortizations data}

\item{emprestimo}{A vector with value of loan for each period}

\item{num_round}{number to round. If NA, no rounding is done.}
}
\value{
balance of long term loan after amortization
}
\description{
computes balance between loan and amortiation (for long term loan)
}
\examples{
 gera_saldo_fin(c(0, 50, 50), c(100, 0, 0))

}
