% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_saldo_ponte.R
\name{gera_saldo_ponte}
\alias{gera_saldo_ponte}
\title{Saldo para emprestimo ponte}
\usage{
gera_saldo_ponte(amortizacao, emprestimo)
}
\arguments{
\item{amortizacao}{A vector with amortizations data}

\item{emprestimo}{A vector with value of loan (repeat total value)}
}
\value{
balance of 'emprestimo ponte' after amortization
}
\description{
computes balance between loan and amortiation (for 'emprestimo ponte')
}
\examples{
 gera_saldo_ponte(c(0, 100, 0), c(100, 100, 100))

}
