% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random_walk}
\alias{random_walk}
\title{Simulates a random walk}
\usage{
random_walk(start_seed, num_years = 33, mu = 0, sd = 0.05)
}
\arguments{
\item{start_seed}{A number}

\item{num_years}{A number}

\item{mu}{A number}

\item{sd}{A number}
}
\value{
a vector of simulated random walk times series of size equal to num_years
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 random_walk(start_seed = 1.3, mu = 0, sd = .05)

}
