% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk_passengers.R
\name{random_walk_passengers}
\alias{random_walk_passengers}
\title{Simulates a random walk for passengers number}
\usage{
random_walk_passengers(start_seed = 196860, num_years = 33, mu = 0,
  sd = 0.7 * start_seed, beta = 1, jump = T)
}
\arguments{
\item{start_seed}{A number. First point of the series.}

\item{num_years}{A number. Number of years. Usually 33.}

\item{mu}{A number}

\item{sd}{A number}

\item{beta}{A number. Effect of past step on the next step.}

\item{jump}{Logical. If TRUE, in the eight year of the series the forecasted demand increases by 25 percent relative to last value (plus error). Equivalent to set beta =1.25 for year 8.}
}
\value{
a vector of simulated random walk times series of size equal to num_years
}
\description{
Auxiliar function. It computes a random walk with jumps, that is used to generate forecasted demand.
}
\examples{
 random_walk_passengers(start_seed=19686)

}
