% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_igpm.R
\name{sim_igpm}
\alias{sim_igpm}
\title{Simulates fiscal impact of line 4 subway PPP}
\usage{
sim_igpm(igpm_0 = 1.1, igpm_realizado)
}
\arguments{
\item{igpm_0}{A number}

\item{igpm_realizado}{A vector or NA}
}
\value{
A vector of inflation for 33 years
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 sim_igpm(igpm_0= 1.1, igpm_realizado = NA)

}
