% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sensibilidade_line4.R
\name{sim_sensibilidade_line4}
\alias{sim_sensibilidade_line4}
\title{Run Monte Carlo simulation for line 4 - sensibilidade parameter}
\usage{
sim_sensibilidade_line4(n_sim, num_years, t_0 = 2.14, ipc_0 = 1.1,
  ipgm_0 = 1.1, a = 0.5, b = 0.5, ipc_realizado = NA,
  igpm_realizado = NA, qualityAdjustment = 1, ajuste_inflacao,
  mu_sense = 1, sd_sense = 0.1, incidencia = c(0, 0, 0, 0.5, rep(1,
  9), rep(0, num_years - 13)), type = "white_noise", start_seed,
  use_random_walk = F, start_value = 196860, mu = start_value,
  sd = 0.07 * start_value, beta = 1, jump = F)
}
\arguments{
\item{n_sim}{A number}

\item{num_years}{A number}

\item{t_0}{A number}

\item{ipc_0}{A number}

\item{ipgm_0}{A number}

\item{a}{A number}

\item{b}{A number}

\item{ipc_realizado}{A number}

\item{igpm_realizado}{A number}

\item{qualityAdjustment}{A number}

\item{ajuste_inflacao}{logical}

\item{mu_sense}{A number}

\item{sd_sense}{A number}

\item{incidencia}{A vector. line 27 of excel.}

\item{type}{character. Either "white_noise", or "random_walk" or "constant"}

\item{start_seed}{A number}

\item{use_random_walk}{Logical. If TRUE, it will use a random walk to forecast demand (instead of a fixed forecast as in excel)}

\item{start_value}{A number. First point of the series. Use only if random_walk = T}

\item{mu}{A number}

\item{sd}{A number}

\item{beta}{A number. Effect of past step on the next step.}

\item{jump}{logical. If TRUE, in the eight year of the series the forecasted demand increases by 25\% relative to last value (plus error). Equivalent to set beta =1.25 for year 8.}
}
\value{
a list with two elements: a data.frame wih simulated difference betwen real demand and forecasted, and a matrizwith sensibilidade parameters generated in the simulation.
}
\description{
It computes: diferença entre compensação (MD) e Pagamento (recebimento) de (para) poder concedente (linha 28 excel). Ou seja, a diferença entre receita prevista em contrato e realizada da simulação.
}
\examples{
 sim_sensibilidade_line4(n_sim=3, num_years=33, start_seed = 1, ajuste_inflacao = F)

}
