% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bid.R
\name{theme_bid}
\alias{theme_bid}
\title{Our ggplot2 theme}
\usage{
theme_bid(tam_fonte = 14, fonte = "sans")
}
\arguments{
\item{tam_fonte}{Set font size used in the plot.}

\item{fonte}{Set main font used in the plot.}
}
\description{
\code{theme_bid} applies our ggplot2 theme to a ggplot2 plot object.
}
