% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_noise.R
\name{white_noise}
\alias{white_noise}
\title{Simulates white noise}
\usage{
white_noise(start_seed, num_years = 33, mu, sd)
}
\arguments{
\item{start_seed}{A number}

\item{num_years}{A number}

\item{mu}{A number}

\item{sd}{A number}
}
\value{
a vector of simulated white noise times series of size equal to num_years
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 white_noise(start_seed = 1.3, mu = 0, sd = .05)

}
