% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aditivo3.R
\name{aditivo3}
\alias{aditivo3}
\title{Multa por Atraso do Aditivo 3}
\usage{
aditivo3(num_meses_atraso, vec_meses_atrasados)
}
\arguments{
\item{num_meses_atraso}{a number - number of months that phase 2 is delayed (asinoriginal contract)}

\item{vec_meses_atrasados}{A vector with numbers - number of delay (in months) for each tranche. There are three tranches.}
}
\value{
total payment for the informed number of months delayed
}
\description{
Computes the fine due to delay in construction, as per in "aditivo 3"
}
\examples{
 aditivo3(num_meses_atraso = 12,
vec_meses_atrasados= c(10, 14, NA))

}
