% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ajuste_demanda.R
\name{gen_ajuste_demanda}
\alias{gen_ajuste_demanda}
\title{Simulates fiscal impact of line 4 subway PPP}
\usage{
gen_ajuste_demanda(sensibilidade = 1, num_years, t_0 = 2.14, ipc_0,
  ipgm_0, a, b, ipc_realizado = NA, igpm_realizado = NA,
  ajuste_inflacao = F)
}
\arguments{
\item{sensibilidade}{A vector}

\item{num_years}{A number}

\item{t_0}{A number}

\item{ipc_0}{A number}

\item{ipgm_0}{A number}

\item{a}{A number}

\item{b}{A number}

\item{ipc_realizado}{A number}

\item{igpm_realizado}{A number}

\item{ajuste_inflacao}{logical}
}
\value{
a vector of demand adjustement of size equal to num_years - input to compute implicit rate
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 gen_num_passengers(sensibilidade = 1.3)

}
