% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_implicit_ticket_revenue.R
\name{gen_implicit_ticket_revenue}
\alias{gen_implicit_ticket_revenue}
\title{Gera tarifa implicita da linha 4}
\usage{
gen_implicit_ticket_revenue(numPassengersExclusive,
  numPassengersIntegrated, price_ticket, num_years, t_0 = 2.14,
  ipc_0 = 1.1, ipgm_0 = 1.1, a = 0.5, b = 0.5,
  ipc_realizado = NA, igpm_realizado = NA, sensibilidade,
  ajuste_inflacao)
}
\arguments{
\item{numPassengersExclusive}{A vector}

\item{numPassengersIntegrated}{A vector}

\item{price_ticket}{A vector}

\item{num_years}{A number}

\item{t_0}{A number}

\item{ipc_0}{A number}

\item{ipgm_0}{A number}

\item{a}{A number}

\item{b}{A number}

\item{ipc_realizado}{A number}

\item{igpm_realizado}{A number}

\item{sensibilidade}{A vector}

\item{ajuste_inflacao}{logical}
}
\value{
a vector of demand adjustement of size equal to num_years - input to compute implicit rate
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 gen_num_passengers(sensibilidade = 1.3)

}
