% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_num_passengers.R
\name{gen_num_passengers}
\alias{gen_num_passengers}
\title{Generates passengers for line 4}
\usage{
gen_num_passengers(sensibilidade = 1, num_years = 33)
}
\arguments{
\item{sensibilidade}{A vector}

\item{num_years}{A number}
}
\value{
A list of four elements, all with size 33: numPassengersExclusive, numPassengersIntegrated, demanda_projetada and demanda_real
}
\description{
It generate a forecast of passengers for all years of contract, with integrated and exclusie passengers.
}
\examples{
 gen_num_passengers(sensibilidade = 1.3)

}
