% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_num_passengers.R
\name{gen_num_passengers}
\alias{gen_num_passengers}
\title{Simulates fiscal impact of line 4 subway PPP}
\usage{
gen_num_passengers(sensibilidade = 1)
}
\arguments{
\item{sensibilidade}{A vector}
}
\value{
A list of four elements, all with size 33: numPassengersExclusive, numPassengersIntegrated, demanda_projetada and demanda_real
}
\description{
This package allows the user to run Monte Carlo simulation to assess the fiscal impact of lline 4 PPP in São Paulo.
}
\examples{
 gen_num_passengers(sensibilidade = 1.3)

}
