% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R, R/random_walk_passengers.R
\name{random_walk}
\alias{random_walk}
\title{Simulates a random walk}
\usage{
random_walk(start_seed = 196860, num_years = 33, mu = start_seed,
  sd = 0.7 * start_seed, beta = 1, jump = T)

random_walk(start_seed = 196860, num_years = 33, mu = start_seed,
  sd = 0.7 * start_seed, beta = 1, jump = T)
}
\arguments{
\item{start_seed}{A number. First point of the series.}

\item{num_years}{A number}

\item{mu}{A number}

\item{sd}{A number}

\item{beta}{A number. Effect of past step on the next step.}

\item{jump}{logica. If TRUE, in the eight year of the series the forecasted demand increases by 25% relative to last value (plus error). Equivalent to set beta =1.25 for year 8.}

\item{start_seed}{A number. First point of the series.}

\item{num_years}{A number}

\item{mu}{A number}

\item{sd}{A number}

\item{beta}{A number. Effect of past step on the next step.}
}
\value{
a vector of simulated random walk times series of size equal to num_years

a vector of simulated random walk times series of size equal to num_years
}
\description{
Auxiliar function. It computes a wanrdom walk, that is used to generate sensibility. In other words, its results are used as themultiplier of real demand. A value os 1.1 means for instance that the demand is 1.1 bigger than projected.

Auxiliar function. It computes a random walk with jumps, that is used to generate forecasted demand.
}
\examples{
 random_walk(start_seed = 1.3, mu = 0, sd = .05)

 random_walk(start_seed = 196860)

}
