% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sensibilidade_line4.R
\name{sim_sensibilidade_line4}
\alias{sim_sensibilidade_line4}
\title{Run Monte Carlo simulation for line 4 - sensibilidade parameter}
\usage{
sim_sensibilidade_line4(n_sim, num_years, t_0 = 2.14, ipc_0 = 1.1,
  ipgm_0 = 1.1, a = 0.5, b = 0.5, ipc_realizado = NA,
  igpm_realizado = NA, qualityAdjustment = 1, ajuste_inflacao,
  mu = 1, sd = 0.1, incidencia = c(0, 0, 0, 0.5, rep(1, 9), rep(0,
  num_years - 13)), type = "white_noise", start_seed)
}
\arguments{
\item{n_sim}{A number}

\item{num_years}{A number}

\item{t_0}{A number}

\item{ipc_0}{A number}

\item{ipgm_0}{A number}

\item{a}{A number}

\item{b}{A number}

\item{ipc_realizado}{A number}

\item{igpm_realizado}{A number}

\item{qualityAdjustment}{A number}

\item{ajuste_inflacao}{logical}

\item{mu}{A number}

\item{sd}{A number}

\item{incidencia}{A vector. line 27 of excel.}

\item{type}{character. Either "white_noise", or "random_walk" or "constant"}

\item{start_seed}{A number}
}
\value{
a list with two elements: a data.frame wih simulated difference betwen real demand and forecasted, and a matrizwith sensibilidade parameters generated in the simulation.
}
\description{
It computes: diferença entre compensação (MD) e Pagamento (recebimento) de (para) poder concedente (linha 28 excel). Ou seja, a diferença entre receita prevista em contrato e realizada da simulação.
}
\examples{
 sim_sensibilidade_line4(3, 33, start_seed = 1, ajuste_inflacao = F)

}
